added folders to security
