+++ src/main/java/theChillys/chillys_radio/security/role/IRoleService.java
@@ -1,4 +1,4 @@
-package theChillys.chillys_radio.security;
+package theChillys.chillys_radio.security.role;
 
 public interface IRoleService {
     Role getRoleByTitle(String title);
