+++ src/main/java/theChillys/chillys_radio/security/role/RoleServiceImpl.java
@@ -1,4 +1,4 @@
-package theChillys.chillys_radio.security;
+package theChillys.chillys_radio.security.role;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
