+++ src/main/java/theChillys/chillys_radio/security/user/User.java
@@ -1,4 +1,4 @@
-package theChillys.chillys_radio.security;
+package theChillys.chillys_radio.security.user;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
+import theChillys.chillys_radio.security.role.Role;
 
 import java.util.Collection;
 import java.util.Set;
