+++ src/main/java/theChillys/chillys_radio/security/user/UserResponseDto.java
@@ -1,17 +1,20 @@
-package theChillys.chillys_radio.security;
+package theChillys.chillys_radio.security.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import theChillys.chillys_radio.security.role.Role;
+
+import java.util.Set;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
-public class UserRequestDto {
-
+public class UserResponseDto {
+    private Long id;
     private String name;
-    private String password;
     private String email;
+    private Set<Role> roles;
 }
